requirements.txt
certifi==2025.8.3
charset-normalizer==2.0.12
click==8.2.1
Flask==2.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.0.33
numpy==1.26.4
pandas==1.5.3
protobuf==3.20.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.27.1
six==1.17.0
tqdm==4.65.0
urllib3==1.26.20
waitress==2.1.2
Werkzeug==2.2.2
